# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Hub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /connections:
    get:
      operationId: listConnections
      tags: [connections]
      summary: List all platform connections
      responses:
        "200":
          description: List of connections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Connection"
    post:
      operationId: createConnection
      tags: [connections]
      summary: Create or update a platform connection (API key/token based)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConnectionBody"
      responses:
        "201":
          description: Created connection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
  /connections/{id}:
    delete:
      operationId: deleteConnection
      tags: [connections]
      summary: Disconnect/delete a platform connection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /connections/{id}/test:
    post:
      operationId: testConnection
      tags: [connections]
      summary: Test a platform connection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionTestResult"
  /connections/oauth/{platform}/initiate:
    get:
      operationId: initiateOAuth
      tags: [connections]
      summary: Start OAuth flow for a platform
      parameters:
        - name: platform
          in: path
          required: true
          schema:
            type: string
            enum: [linkedin, google, twitter, meta]
      responses:
        "200":
          description: OAuth URL to redirect user to
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthInitiateResponse"
  /connections/oauth/{platform}/callback:
    get:
      operationId: oauthCallback
      tags: [connections]
      summary: OAuth callback handler
      parameters:
        - name: platform
          in: path
          required: true
          schema:
            type: string
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: error
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect after OAuth

  # ─── Connection Actions ───────────────────────────────────────────────────
  /connections/{id}/actions/post:
    post:
      operationId: connectionPost
      tags: [connections]
      summary: Post content via a connected platform
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectionPostBody"
      responses:
        "200":
          description: Post result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionActionResult"
  /connections/{id}/actions/send-message:
    post:
      operationId: connectionSendMessage
      tags: [connections]
      summary: Send a DM or message via a connected platform
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectionSendMessageBody"
      responses:
        "200":
          description: Message result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionActionResult"
components:
  schemas:
...
    Connection:
      type: object
      properties:
        id:
          type: integer
        platform:
          type: string
          enum: [linkedin, google, twitter, meta, gohighlevel, email]
        displayName:
          type: string
        accountLabel:
          type: string
          nullable: true
        isConnected:
          type: boolean
        authType:
          type: string
          enum: [oauth, api_key]
        scopes:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - platform
        - displayName
        - isConnected
        - authType
        - scopes
        - createdAt
    CreateConnectionBody:
      type: object
      properties:
        platform:
          type: string
          enum: [gohighlevel, email]
        apiKey:
          type: string
        accountLabel:
          type: string
        metadata:
          type: object
      required:
        - platform
        - apiKey
    ConnectionTestResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        accountInfo:
          type: object
          nullable: true
      required:
        - success
        - message
    OAuthInitiateResponse:
      type: object
      properties:
        url:
          type: string
        state:
          type: string
      required:
        - url
        - state
    ConnectionPostBody:
      type: object
      properties:
        content:
          type: string
        mediaUrls:
          type: array
          items:
            type: string
        targetId:
          type: string
          nullable: true
      required:
        - content
    ConnectionSendMessageBody:
      type: object
      properties:
        recipientId:
          type: string
        message:
          type: string
      required:
        - recipientId
        - message
    ConnectionActionResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        postId:
          type: string
          nullable: true
      required:
        - success
        - message