# lib/api-zod/src/generated/types/connection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Hub API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ConnectionAuthType } from "./connectionAuthType";
import type { ConnectionPlatform } from "./connectionPlatform";

export interface Connection {
  id: number;
  platform: ConnectionPlatform;
  displayName: string;
  accountLabel?: string | null;
  isConnected: boolean;
  authType: ConnectionAuthType;
  scopes: string[];
  createdAt: Date;
  expiresAt?: Date | null;
}