# lib/api-zod/src/generated/types/connectionActionResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Hub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ConnectionActionResult {
  success: boolean;
  message: string;
  postId?: string | null;
}