# lib/api-zod/src/generated/types/connectionPostBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Hub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ConnectionPostBody {
  content: string;
  mediaUrls?: string[];
  targetId?: string | null;
}